+++ src/Builders/RequestBuilder/GpApi/GpApiRecurringRequestBuilder.php
@@ -39,12 +39,16 @@ class GpApiRecurringRequestBuilder implements IRequestBuilder
         /** @var RecurringBuilder $builder */
         $this->builder = $builder;
         $requestData = [];
-        /**
-         * @var RecurringBuilder $builder
-         */
+        /** @var string $merchantUrl */
+        $merchantUrl = '';
+       
+        if($config->merchantId != null){
+            $merchantUrl = "/merchants/".$config->merchantId;
+        }
         switch ($builder->transactionType) {
             case TransactionType::CREATE:
                 $endpoint = GpApiRequest::PAYERS_ENDPOINT;
+                $endpoint = $merchantUrl . GpApiRequest::PAYERS_ENDPOINT;
                 $verb = 'POST';
                 if ($builder->entity instanceof Customer) {
                     $this->preparePayerRequest($requestData);
@@ -52,6 +56,7 @@ class GpApiRecurringRequestBuilder implements IRequestBuilder
                 break;
             case TransactionType::EDIT:
                 $endpoint = GpApiRequest::PAYERS_ENDPOINT . '/' . $this->builder->entity->id;
+                $endpoint = $merchantUrl . GpApiRequest::PAYERS_ENDPOINT . "/";
                 $verb = 'PATCH';
                 if ($builder->entity instanceof Customer) {
                     $this->preparePayerRequest($requestData);