+++ src/Terminals/Enums/Reinitialize.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace GlobalPayments\Api\Terminals\Enums;
+
+use GlobalPayments\Api\Entities\Enum;
+
+class Reinitialize extends Enum
+{
+    const REINITIALIZE_APPLICATION = 1;
+    const NO_REINITIALIZE_APPLICATION = 0;
+}